Implement strict mode for locking
In strict mode, a locked configuration without a lockfile, empty or not,
 is illegal and will cause the build to fail.

Issue #9907
