+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/locking/DependencyLockingIntegrationTest.groovy
@@ -16,236 +16,14 @@
 
 package org.gradle.integtests.resolve.locking
 
-import org.gradle.integtests.fixtures.AbstractDependencyResolutionTest
+import org.gradle.api.artifacts.dsl.LockMode
 import org.gradle.integtests.fixtures.ToBeFixedForInstantExecution
-import org.gradle.integtests.fixtures.resolve.ResolveTestFixture
-import spock.lang.Unroll
 
-class DependencyLockingIntegrationTest extends AbstractDependencyResolutionTest {
+class DependencyLockingIntegrationTest extends AbstractLockingIntegrationTest {
 
-    def lockfileFixture = new LockfileFixture(testDirectory: testDirectory)
-    ResolveTestFixture resolve
-
-    def setup() {
-        settingsFile << "rootProject.name = 'depLock'"
-        resolve = new ResolveTestFixture(buildFile, "lockedConf")
-        resolve.prepare()
-        resolve.addDefaultVariantDerivationStrategy()
-    }
-
-    def 'succeeds when lock file does not conflict from declared versions'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0'])
-
-        when:
-        succeeds 'checkDeps'
-
-        then:
-        resolve.expectDefaultConfiguration('runtime')
-        resolve.expectGraph {
-            root(":", ":depLock:") {
-                edge("org:foo:1.+", "org:foo:1.0")
-                constraint("org:foo:{strictly 1.0}", "org:foo:1.0") {
-                    byConstraint("dependency was locked to version '1.0'")
-                }
-            }
-        }
-    }
-
-    @Unroll
-    def 'fails when lock file conflicts with declared strict constraint'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf('org:foo') {
-        version { strictly '1.1' }
-    }
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0'])
-
-        when:
-        fails 'checkDeps'
-
-        then:
-        failure.assertHasCause """Cannot find a version of 'org:foo' that satisfies the version constraints:
-   Dependency path ':depLock:unspecified' --> 'org:foo:1.+'
-   Dependency path ':depLock:unspecified' --> 'org:foo:{strictly 1.1}'
-   Constraint path ':depLock:unspecified' --> 'org:foo:{strictly 1.0}' because of the following reason: dependency was locked to version '1.0'"""
-    }
-
-    def 'fails when lock file conflicts with declared version constraint'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf('org:foo:1.1')
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0'])
-
-        when:
-        fails 'checkDeps'
-
-        then:
-        failure.assertHasCause """Cannot find a version of 'org:foo' that satisfies the version constraints:
-   Dependency path ':depLock:unspecified' --> 'org:foo:1.+'
-   Dependency path ':depLock:unspecified' --> 'org:foo:1.1'
-   Constraint path ':depLock:unspecified' --> 'org:foo:{strictly 1.0}' because of the following reason: dependency was locked to version '1.0'"""
-    }
-
-    def 'fails when lock file contains entry that is not in resolution result'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:bar:1.0', 'org:foo:1.0', 'org:baz:1.0'])
-
-        when:
-        fails 'checkDeps'
-
-        then:
-        failure.assertHasCause("Could not resolve all dependencies for configuration ':lockedConf'.")
-        failure.assertHasCause("Did not resolve 'org:bar:1.0' which is part of the dependency lock state")
-        failure.assertHasCause("Did not resolve 'org:baz:1.0' which is part of the dependency lock state")
-    }
-
-    def 'fails when lock file does not contain entry for module in resolution result'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf 'org:bar:1.+'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0'])
-
-        when:
-        fails 'checkDeps'
-
-        then:
-        failure.assertHasCause("Could not resolve all dependencies for configuration ':lockedConf'.")
-        failure.assertHasCause("Resolved 'org:bar:1.0' which is not part of the dependency lock state")
-    }
-
-    def 'fails when resolution result is empty and lock file contains entries'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-"""
-        lockfileFixture.createLockfile('lockedConf', ['org:foo:1.0'])
-
-        when:
-        fails 'checkDeps'
-
-        then:
-        failure.assertHasCause('Could not resolve all dependencies for configuration \':lockedConf\'.')
-        failure.assertHasCause('Did not resolve \'org:foo:1.0\' which is part of the dependency lock state')
+    @Override
+    LockMode lockMode() {
+        LockMode.DEFAULT
     }
 
     @ToBeFixedForInstantExecution
@@ -277,195 +55,6 @@ dependencies {
 
         then:
         lockfileFixture.expectMissing('unlockedConf')
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'does not write-locks for unlocked configuration'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-
-        buildFile << """
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    unlockedConf
-}
-
-dependencies {
-    unlockedConf 'org:foo:1.+'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.expectMissing('unlockedConf')
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'dependency report passes with failed dependencies using out-of-date lock file'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    constraints {
-        lockedConf('org:foo:1.1')
-    }
-    lockedConf 'org:foo:1.+'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0'])
-
-        when:
-        run 'dependencies'
-
-        then:
-        outputContains """lockedConf
-+--- org:foo:1.+ FAILED
-+--- org:foo:1.1 FAILED
-\\--- org:foo:{strictly 1.0} FAILED"""
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'dependency report passes with FAILED dependencies for all out lock issues'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    constraints {
-        lockedConf('org:foo:1.1')
-    }
-    lockedConf 'org:foo:[1.0, 1.1]'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:bar:1.0', 'org:foo:1.0'])
-
-        when:
-        run 'dependencies'
-
-        then:
-        outputContains """lockedConf
-+--- org:foo:[1.0, 1.1] FAILED
-+--- org:foo:1.1 FAILED
-+--- org:foo:{strictly 1.0} FAILED
-\\--- org:bar:1.0 FAILED
-"""
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'writes dependency lock file when requested'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf 'org:bar:1.+'
-}
-"""
-
-        when:
-        succeeds'dependencies', '--write-locks', '--refresh-dependencies'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0'])
-
-    }
-
-    @Unroll
-    @ToBeFixedForInstantExecution
-    def "writes dependency lock file for resolved version #version"() {
-        mavenRepo.module('org', 'bar', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.1').publish()
-        mavenRepo.module('org', 'bar', '2.0').publish()
-        mavenRepo.module('org', 'bar', '2.1').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-    subConf
-}
-
-dependencies {
-    lockedConf 'org:bar:${version}'
-    subConf 'org:bar:1.1'
-}
-"""
-
-        when:
-        succeeds'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ["org:bar:${resolved}"])
-
-        where:
-        version     | resolved
-        "[1.0,)"    | "2.1"
-        "[1.0,2.0)" | "1.1"
-        "[1.0,2.0]" | "2.0"
-        "(,2.0)"    | "1.1"
-        "1.+"       | "1.1"
-        "+"         | "2.1"
     }
 
     def "version selector combinations are resolved equally for locked and unlocked configurations"() {
@@ -514,348 +103,6 @@ task check {
     }
 
     @ToBeFixedForInstantExecution
-    def "does not lock a configuration that is marked with deactivateDependencyLocking"() {
-        ['foo', 'foz', 'bar', 'baz'].each { artifact ->
-            mavenRepo.module('org', artifact, '1.0').publish()
-            mavenRepo.module('org', artifact, '1.1').publish()
-            mavenRepo.module('org', artifact, '1.2').publish()
-            mavenRepo.module('org', artifact, '2.0').publish()
-        }
-
-        buildFile << """
-repositories {
-    maven {
-        url '${mavenRepo.uri}'
-    }
-}
-
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-configurations {
-    conf
-    lockEnabledConf {
-        extendsFrom conf
-    }
-     secondLockEnabledConf {
-        extendsFrom lockEnabledConf
-        resolutionStrategy.deactivateDependencyLocking()
-    }
-}
-dependencies {
-    conf 'org:foo:[1.0,)'
-    conf 'org:foo:1.1'
-
-    conf 'org:foz:latest.integration'
-    conf 'org:foz:1.1'
-
-    conf 'org:bar:1.+'
-    conf 'org:bar:1.1'
-
-    conf 'org:baz:+'
-    conf 'org:baz:1.1'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockEnabledConf', ['org:bar:1.2', 'org:baz:2.0', 'org:foo:1.1', 'org:foz:2.0'])
-        lockfileFixture.expectMissing('secondLockEnabledConf')
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'upgrades lock file'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-        mavenRepo.module('org', 'foo', '2.0').publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-}
-"""
-
-
-        lockfileFixture.createLockfile('lockedConf', ["org:foo:1.0"])
-
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.1'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'counts dependencies with multiple paths as one instance'() {
-        def foo = mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.0').dependsOn(foo).publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf 'org:bar:1.+'
-}
-"""
-
-        lockfileFixture.createLockfile('lockedConf',['org:foo:1.0', 'org:bar:1.0'])
-
-        when:
-        succeeds 'dependencies'
-
-        then:
-        outputContains("org:foo:1.0")
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'does not write duplicates in the lockfile'() {
-        def foo = mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.0').dependsOn(foo).publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-    lockedConf 'org:bar:1.+'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'includes transitive dependencies in the lock file'() {
-        def dep = mavenRepo.module('org', 'bar', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.0').dependsOn(dep).publish()
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:1.+'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--configuration', 'lockedConf', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'updates part of the lockfile'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.1').publish()
-
-        lockfileFixture.createLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0'])
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:[1.0,2.0)'
-    lockedConf 'org:bar:[1.0,2.0)'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--update-locks', 'org:foo'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.1', 'org:bar:1.0'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'updates part of the lockfile using wildcard'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-        mavenRepo.module('org', 'bar', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.1').publish()
-
-        lockfileFixture.createLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0'])
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:[1.0,2.0)'
-    lockedConf 'org:bar:[1.0,2.0)'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--update-locks', 'org:f*'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.1', 'org:bar:1.0'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'updates but ignores irrelevant modules'() {
-        mavenRepo.module('org', 'bar', '1.0').publish()
-        mavenRepo.module('org', 'bar', '1.1').publish()
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-
-        lockfileFixture.createLockfile('lockedConf', ['org:bar:1.0', 'org:foo:1.0'])
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:bar:[1.0,2.0)'
-    lockedConf 'org:foo:[1.0,2.0)'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--update-locks', 'org:foo,org:baz'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:bar:1.0', 'org:foo:1.1'])
-    }
-
-    @ToBeFixedForInstantExecution
-    def 'updates multiple parts of the lockfile'() {
-        mavenRepo.module('org', 'foo', '1.0').publish()
-        mavenRepo.module('org', 'foo', '1.1').publish()
-        mavenRepo.module('org', 'buz', '1.0').publish()
-        mavenRepo.module('org', 'buz', '1.1').publish()
-        def bar10 = mavenRepo.module('org', 'bar', '1.0').publish()
-        def bar11 = mavenRepo.module('org', 'bar', '1.1').publish()
-        mavenRepo.module('org', 'baz', '1.0').dependsOn(bar10).publish()
-        mavenRepo.module('org', 'baz', '1.1').dependsOn(bar11).publish()
-
-        lockfileFixture.createLockfile('lockedConf', ['org:foo:1.0', 'org:bar:1.0', 'org:baz:1.0', 'org:buz:1.0'])
-
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-
-dependencies {
-    lockedConf 'org:foo:[1.0,2.0)'
-    lockedConf 'org:baz:[1.0,2.0)'
-    lockedConf 'org:buz:[1.0,2.0)'
-}
-"""
-
-        when:
-        succeeds 'dependencies', '--update-locks', 'org:foo,org:baz'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', ['org:foo:1.1', 'org:bar:1.1', 'org:baz:1.1', 'org:buz:1.0'])
-    }
-
-    @ToBeFixedForInstantExecution
     def 'writes a new lock file if update done without lockfile present'() {
         mavenRepo.module('org', 'foo', '1.0').publish()
         mavenRepo.module('org', 'foo', '1.1').publish()
@@ -891,30 +138,6 @@ dependencies {
     }
 
     @ToBeFixedForInstantExecution
-    def 'writes an empty lock file for an empty configuration'() {
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-"""
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', [])
-    }
-
-    @ToBeFixedForInstantExecution
     def 'does not write an empty lock file for an empty configuration if not requested'() {
         buildFile << """
 dependencyLocking {
@@ -938,63 +161,4 @@ configurations {
         lockfileFixture.expectMissing('lockedConf')
     }
 
-    @ToBeFixedForInstantExecution
-    def 'overwrites a not empty lock file with an empty one when configuration no longer has dependencies'() {
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-repositories {
-    maven {
-        name 'repo'
-        url '${mavenRepo.uri}'
-    }
-}
-configurations {
-    lockedConf
-}
-"""
-        lockfileFixture.createLockfile('lockedConf', ['org:foo:1.0'])
-
-        when:
-        succeeds 'dependencies', '--write-locks'
-
-        then:
-        lockfileFixture.verifyLockfile('lockedConf', [])
-    }
-
-    @Unroll
-    @ToBeFixedForInstantExecution
-    def "fails if trying to resolve a locked configuration with #flag"() {
-        buildFile << """
-dependencyLocking {
-    lockAllConfigurations()
-}
-
-configurations {
-    lockedConf {
-        resolutionStrategy {
-            $flag()
-        }
-    }
-}
-
-dependencies {
-    lockedConf 'org:foo:1.0'
-}
-"""
-
-        when:
-        fails 'dependencies'
-
-        then:
-        failure.assertHasCause "Resolution strategy has both dependency locking and fail on $desc versions enabled. You must choose between the two modes."
-
-        where:
-        flag                              | desc
-        'failOnDynamicVersions'           | 'dynamic'
-        'failOnChangingVersions'          | 'changing'
-        'failOnNonReproducibleResolution' | 'dynamic'
-    }
 }
