+++ subprojects/dependency-management/src/main/java/org/gradle/internal/locking/DefaultDependencyLockingState.java
@@ -17,7 +17,7 @@
 package org.gradle.internal.locking;
 
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
-import org.gradle.api.artifacts.dsl.DependencyLockingHandler;
+import org.gradle.api.artifacts.dsl.LockMode;
 import org.gradle.api.internal.artifacts.dsl.dependencies.DependencyLockingState;
 
 import java.util.Set;
@@ -30,14 +30,14 @@ public class DefaultDependencyLockingState implements DependencyLockingState {
 
     private final boolean strictlyValidate;
     private final Set<ModuleComponentIdentifier> constraints;
-    private final DependencyLockingHandler.LockMode lockMode;
+    private final LockMode lockMode;
 
     private DefaultDependencyLockingState() {
         strictlyValidate = false;
         constraints = emptySet();
-        lockMode = DependencyLockingHandler.LockMode.DEFAULT;
+        lockMode = LockMode.DEFAULT;
     }
-    public DefaultDependencyLockingState(boolean partialUpdate, Set<ModuleComponentIdentifier> constraints, DependencyLockingHandler.LockMode lockMode) {
+    public DefaultDependencyLockingState(boolean partialUpdate, Set<ModuleComponentIdentifier> constraints, LockMode lockMode) {
         strictlyValidate = !partialUpdate;
         this.constraints = constraints;
         this.lockMode = lockMode;
