+++ subprojects/dependency-management/src/main/java/org/gradle/internal/locking/NoOpDependencyLockingProvider.java
@@ -17,7 +17,7 @@
 package org.gradle.internal.locking;
 
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
-import org.gradle.api.artifacts.dsl.DependencyLockingHandler;
+import org.gradle.api.artifacts.dsl.LockMode;
 import org.gradle.api.internal.artifacts.dsl.dependencies.DependencyLockingProvider;
 import org.gradle.api.internal.artifacts.dsl.dependencies.DependencyLockingState;
 
@@ -46,12 +46,12 @@ public class NoOpDependencyLockingProvider implements DependencyLockingProvider 
     }
 
     @Override
-    public DependencyLockingHandler.LockMode getLockMode() {
+    public LockMode getLockMode() {
         throw new IllegalStateException("Should not be invoked on the no-op instance");
     }
 
     @Override
-    public void setLockMode(DependencyLockingHandler.LockMode mode) {
+    public void setLockMode(LockMode mode) {
         throw new IllegalStateException("Should not be invoked on the no-op instance");
     }
 }
